// src/requests/enum/request-type.enum.ts
export enum RequestType {
  PRODUCT_CREATE = 'PRODUCT_CREATE',
  PRODUCT_UPDATE = 'PRODUCT_UPDATE',

  VARIANT_CREATE = 'VARIANT_CREATE',
  VARIANT_UPDATE = 'VARIANT_UPDATE',

  STOCK_IN = 'STOCK_IN',
  STOCK_OUT = 'STOCK_OUT',
  STOCK_TRANSFER = 'STOCK_TRANSFER',
  STOCK_ADJUSTMENT = 'STOCK_ADJUSTMENT',
}